refactor: centralize validator types
