+++ tools/template-validator/src/configurations.ts
@@ -1,5 +1,5 @@
 // tools/template-validator/src/configurations.ts
-import { TemplateConfig } from "./types";
+import type { TemplateConfig } from "@farm/types";
 
 export const AI_CHAT_CONFIGURATIONS: TemplateConfig[] = [
   {
